+++ viewer/packages/360-images/src/collection/DefaultImage360Collection.ts
@@ -7,7 +7,7 @@ import pull from 'lodash/pull';
 import { Image360Collection } from './Image360Collection';
 import { Image360Entity } from '../entity/Image360Entity';
 import { Image360EnteredDelegate, Image360ExitedDelegate } from '../types';
-import { Image360CollectionIcons } from '../icons/Image360CollectionIcons';
+import { Image360CollectionIcons } from '../icons/IconCollection';
 
 type Image360Events = 'image360Entered' | 'image360Exited';
