+++ viewer/packages/360-images/src/collection/Image360CollectionFactory.ts
@@ -8,7 +8,7 @@ import zip from 'lodash/zip';
 import { DefaultImage360Collection } from './DefaultImage360Collection';
 import { Image360Entity } from '../entity/Image360Entity';
 import { Image360Icon } from '../icons/Image360Icon';
-import { Image360CollectionIcons } from '../icons/Image360CollectionIcons';
+import { Image360CollectionIcons } from '../icons/IconCollection';
 
 export class Image360CollectionFactory<T> {
   private readonly _image360DataProvider: Image360Provider<T>;
