+++ viewer/packages/360-images/src/icons/IconCollection.ts
